Analytics/requirements.txt
mysql-connector-python
pymongo
datetime
